OS-7.5: Added JS in search page, start tests in ProductService

+++ OnlineStore_Epam2018/SA.OnlineStore.Bussines/Service/Implementation/OrderService.cs
@@ -21,7 +21,10 @@
 
         public void DeleteOrderByOrderId(int Id)
         {
-            _orderRepository.Delete(Id);
+            if (Id > 0)
+            {
+                _orderRepository.Delete(Id);
+            }
         }
 
         public Order GetOrder(int Id)
@@ -48,7 +51,10 @@
 
         public void SaveDefaultOrder(int id)
         {
-             _orderRepository.SaveDefaultOrder(id);
+            if (id > 0)
+            {
+                _orderRepository.SaveDefaultOrder(id);
+            }
         }
 
         public void EditOrder(Order model)
